Move to lazy init and separate data
